Fix gl context error on win10/intel UHD
Avoid unnecessary calls to `glfwMakeContextCurrent` in PointViz that
print errors on windows/intel UHD. Didn't exactly find the root cause,
but a few remarks online suggest this is a driver bug.

This MR also generally improves how the GL context is initialized by
unbinding it after initialization. That way, the render loop can be
safely started in a thread other than the one where PointViz was
instantiated.

Also fix a bunch of warnings under MSVC and enable treating warnings as
errors in CI.

+++ python/src/cpp/_client.cpp
@@ -603,7 +603,7 @@ PYBIND11_PLUGIN(_client) {
             "_complete",
             [](const LidarScan& self,
                nonstd::optional<sensor::AzimuthWindow> window) {
-                if (!window) window = {0, self.w - 1};
+                if (!window) window = {0, static_cast<int>(self.w - 1)};
 
                 const auto& status = self.status();
                 auto start = window.value().first;
